+++ assets/tvs/multitv/configs/links.config.inc.php
@@ -76,7 +76,7 @@ $settings['configuration'] = array(
 	'csvseparator' => ','
 );
 $settings['templatesTest'] = array(
-	'outerTpl' => '<ul>[+wrapper+]</ul>',
+	'outerTpl' => '<ul>[+wrapper+][+pagination+]</ul>',
 	'rowTpl' => '<li>[+text+], [+image+], [+thumb+], [+textarea+], [+date+], [+dropdown+], [+listbox+], [+listbox-multiple+], [+checkbox+], [+option+]</li>'
 		)
 ?>