+++ assets/tvs/multitv/languages/svenska-utf8.language.inc.php
@@ -1,0 +1,49 @@
+<?php
+$language['editSwitch'] = 'Innehåll ärvt från förälder och/eller borttaget. Klicka <a href="#">här</a> för att redigera.';
+$language['clear'] = 'Ta bort alla element';
+$language['confirmclear'] = 'Vill du verkligen ta bort all data?';
+$language['paste'] = 'Klistra in tabelldata (dvs HTML, Word, CSV)';
+$language['pastehere'] = 'Klistra in här:';
+$language['add'] = 'Lägg till';
+$language['remove'] = 'Ta bort';
+$language['edit'] = 'Redigera';
+$language['save'] = 'Spara';
+$language['replace'] = 'Ersätt';
+$language['append'] = 'Bifoga';
+$language['cancel'] = 'Avbryt';
+$language['word'] = 'Word/HTML';
+$language['google'] = 'Google Dokument';
+$language['csv'] = 'CSV';
+$language['paginate.first'] = '&lsaquo; Första';
+$language['paginate.prev'] = '&lt;';
+$language['paginate.next'] = '&gt;';
+$language['paginate.last'] = 'Sista &rsaquo;';
+$language['connector.illegal'] = 'Felaktig parameter!';
+$language['connector.nodoc'] = 'Dokumentet finns inte! Spara ett nytt dokument innan du lägger till en rad.';
+$language['connector.rights'] = 'Otillräckliga rättigheter för denna aktion!';
+$language['connector.nomultitv'] = 'multiTV finns inte!';
+$language['connector.noprocessor'] = 'Processorn finns inte!';
+$language['dataTables'] = '{
+	"sEmptyTable":     "Ingen tillgänglig data i tabellen",
+    "sInfo":           "Visar _START_ till _END_ av _TOTAL_ poster",
+    "sInfoEmpty":      "Visar 0 till 0 av 0 poster",
+    "sInfoFiltered":   "(filtrerad från totalt _MAX_ poster)",
+    "sInfoPostFix":    "",
+    "sInfoThousands":  ",",
+    "sLengthMenu":     "Visa _MENU_ poster",
+    "sLoadingRecords": "Laddar …",
+    "sProcessing":     "Behandlar …",
+    "sSearch":         "Sök:",
+    "sZeroRecords":    "Inga poster hittades",
+    "oPaginate": {
+        "sFirst":    "Första",
+        "sLast":     "Sista",
+        "sNext":     "Nästa",
+        "sPrevious": "Föregående"
+    },
+    "oAria": {
+        "sSortAscending":  ": aktivera för att sortera kolumnen stigande",
+        "sSortDescending": ": aktivera för att sortera kolumnen fallande"
+    }
+}';
+?>